Clippy fixes applied and FPS bug highlighted
Applied several of "cargo clippy" 's suggested fixes, to neaten up the code and avoid redundancy.

There is a bug where the runtime quickly overflows because a variable wasn't being updated. I patched that, which revealed the fundamental bug: miliseconds aren't precise enough as the update function is being called many times per milisecond. I've commented the issue, but not tried to resolve it.